Tickets are now filtered from the controller

+++ app/Http/Controllers/TicketController.php
@@ -5,32 +5,47 @@ namespace App\Http\Controllers;
 use App\Http\Requests\TicketRequest;
 use App\Http\Resources\TicketResource;
 use App\Models\Ticket;
+use Illuminate\Support\Facades\Gate;
 
 class TicketController extends Controller
 {
     public function index()
     {
-        return TicketResource::collection(Ticket::all());
+        $tickets = auth('sanctum')->user()->role === 'admin'
+            ? Ticket::all()
+            : Ticket::where('user_id', auth('sanctum')->user()->id)->get();
+
+        return TicketResource::collection($tickets);
     }
 
     public function store(TicketRequest $request)
     {
-        $ticket = Ticket::create($request->validated());
+        Ticket::create($request->validated());
 
-        $tickets = Ticket::all();
+        $tickets = auth('sanctum')->user()->role === 'admin'
+            ? Ticket::all()
+            : Ticket::where('user_id', auth('sanctum')->user()->id)->get();
+            
         return TicketResource::collection($tickets);
     }
 
     public function update(TicketRequest $request, Ticket $ticket)
     {
+        Gate::authorize('update', $ticket);
+
         $ticket->update($request->validated());
 
-        $tickets = Ticket::all();
+        $tickets = auth('sanctum')->user()->role === 'admin'
+            ? Ticket::all()
+            : Ticket::where('user_id', auth('sanctum')->user()->id)->get();
+            
         return TicketResource::collection($tickets);
     }
 
     public function destroy(Ticket $ticket)
     {
+        Gate::authorize('delete', $ticket);
+
         $ticket->delete();
         
         return response()->json(['message' => 'Ticket succesvol verwijderd']);
